add bootstrap modal dialog

+++ guitar-app/src/app/components/song-sheets/song-sheets.component.ts
@@ -4,6 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { SongSheetsService } from '../../services/song-sheets.service';
 import { SongSheet } from '../../services/song-sheets.model';
 import { RouterLink, RouterOutlet } from '@angular/router';
+import { DialogService } from 'app/services/dialog.service';
 
 @Component({
   selector: 'app-song-sheets',
@@ -16,7 +17,10 @@ export class SongSheetsComponent {
   sheets: SongSheet[] = [];
   newSheetName = '';
 
-  constructor(private service: SongSheetsService) {
+  constructor(
+    private service: SongSheetsService,
+    private dialogService: DialogService
+  ) {
     this.load();
   }
 
@@ -32,8 +36,18 @@ export class SongSheetsComponent {
     }
   }
 
-  deleteSheet(id: string) {
-    this.service.delete(id);
-    this.load();
+  async deleteSheet(id: string) {
+    const confirmed = await this.dialogService.confirm(
+      'Are you sure you want to delete this song sheet?',
+      'Delete Song Sheet',
+      'Delete',
+      'Cancel',
+      { variant: 'danger' }
+    );
+
+    if (confirmed) {
+      this.service.delete(id);
+      this.load();
+    }
   }
 }
